# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Платформа для удаленного управления системой контроля доступа шлагбаумов.",
    "description": "Спецификация API для системы контроля доступа шлагбаумов.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/api",
      "description": "Local docker prod server"
    },
    {
      "url": "http://localhost:8000/api",
      "description": "Local development server"
    },
    {
      "url": "https://gsm-barrier.ru/api",
      "description": "Production API"
    }
  ],
...
  "paths": {
...
    "/barriers/{barrier_id}/": {
      "get": {
        "tags": [
          "Barriers"
        ],
        "summary": "Получение информации о шлагбауме.",
        "description": "Возвращает полную информацию о шлагбауме по его `ID`.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "barrier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 42
            },
            "description": "Уникальный идентификатор шлагбаума."
          }
        ],
        "responses": {
          "200": {
            "description": "Детальная информация о шлагбауме.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarrierView"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный параметр запроса - `barrier_id`."
          },
          "401": {
            "description": "Пользователь не авторизован."
          },
          "403": {
            "description": "Нет доступа к данному шлагбауму."
          },
          "404": {
            "description": "Шлагбаум с указанным `id` не найден."
          },
          "500": {
            "description": "Ошибка сервера. Попробуйте позже."
          }
        }
      }
    },
...
    "/admin/barriers/{barrier_id}/": {
      "get": {
        "tags": [
          "Admin. Barriers."
        ],
        "summary": "Получение данных шлагбаума - для администраторов.",
        "description": "Получение расширенного списка данных для шлагбаума. Только для администратора.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "barrier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 42
            },
            "description": "`ID` удаляемого шлагбаума."
          }
        ],
        "responses": {
          "200": {
            "description": "Детальная информация о шлагбауме.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBarrierView"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный параметр запроса - `barrier_id`."
          },
          "401": {
            "description": "Пользователь не авторизован."
          },
          "403": {
            "description": "Доступ запрещен. Только для администраторов."
          },
          "404": {
            "description": "Шлагбаум с указанным `id` не найден."
          },
          "429": {
            "description": "Превышен лимит запросов. Попробуйте позже."
          },
          "500": {
            "description": "Ошибка сервера. Попробуйте позже."
          }
        }
      },
      "delete": {
        "tags": [
          "Admin. Barriers."
        ],
        "summary": "Удаление шлагбаума (только для администраторов).",
        "description": "Удаление шлагбаума полностью из системы. Только для администратора.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "barrier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 42
            },
            "description": "`ID` удаляемого шлагбаума."
          }
        ],
        "responses": {
          "200": {
            "description": "Шлагбаум успешно удален."
          },
          "400": {
            "description": "Некорректный параметр запроса - `barrier_id`."
          },
          "401": {
            "description": "Пользователь не авторизован."
          },
          "403": {
            "description": "Доступ запрещен. Только для администраторов."
          },
          "404": {
            "description": "Шлагбаум с указанным `id` не найден."
          },
          "429": {
            "description": "Превышен лимит запросов. Попробуйте позже."
          },
          "500": {
            "description": "Ошибка сервера. Попробуйте позже."
          }
        }
      },
      "patch": {
        "tags": [
          "Admin. Barriers."
        ],
        "summary": "Редактирование шлагбаума (только для администраторов).",
        "description": "Позволяет администратору обновить информацию о шлагбауме, включая адрес, телефон устройства, настройки приватности.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "barrier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "example": 42
            },
            "description": "`ID` шлагбаума для обновления."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "additional_info": {
                    "type": "string",
                    "description": "Дополнительная информация о шлагбауме.",
                    "example": "Время работы администратора: По выходным с 12 до 18."
                  },
                  "is_public": {
                    "type": "boolean",
                    "description": "Определяет, можно ли найти шлагбаум через поиск в системе.",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Шлагбаум успешно обновлен.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarrierView"
                }
              }
            }
          },
          "400": {
            "description": "Некорректный параметр запроса - `barrier_id`."
          },
          "401": {
            "description": "Пользователь не авторизован."
          },
          "403": {
            "description": "Нет прав для редактирования данного шлагбаума."
          },
          "404": {
            "description": "Шлагбаум с указанным `id` не найден."
          },
          "429": {
            "description": "Превышен лимит запросов. Попробуйте позже."
          },
          "500": {
            "description": "Ошибка сервера. Попробуйте позже."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BarrierView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 42
          },
          "address": {
            "type": "string",
            "example": "Манежная площадь, 1, Москва"
          },
          "owner": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 42
              },
              "full_name": {
                "type": "string",
                "example": "Ivanov Ivan"
              },
              "phone": {
                "type": "string",
                "example": "+79991234567"
              }
            }
          },
          "device_phone": {
            "type": "string",
            "example": "+79991234567"
          },
          "additional_info": {
            "type": "string",
            "example": "Время работы администратора: По будням с 10 до 18."
          },
          "is_public": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "AdminBarrierView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 42
          },
          "address": {
            "type": "string",
            "example": "Манежная площадь, 1, Москва"
          },
          "owner_id": {
            "type": "integer",
            "example": 42
          },
          "device_phone": {
            "type": "string",
            "example": "+79991234567"
          },
          "device_model": {
            "type": "string",
            "enum": [
              "RTU5025",
              "RTU5035",
              "Telemetrica",
              "Elfoc"
            ],
            "example": "RTU5025"
          },
          "device_phones_amount": {
            "type": "integer",
            "example": 999
          },
          "additional_info": {
            "type": "string",
            "example": "Время работы администратора: По будням с 10 до 18."
          },
          "is_public": {
            "type": "boolean",
            "default": true
          },
          "is_active": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}